refactor: reorganize package structure for Monument-related classes
