+++ monument/src/main/java/pt/estga/monument/controllers/MonumentAdminController.java
@@ -9,7 +9,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import pt.estga.monument.entities.Monument;
-import pt.estga.monument.MonumentMapper;
+import pt.estga.monument.mappers.MonumentMapper;
 import pt.estga.monument.services.MonumentService;
 import pt.estga.monument.dtos.MonumentDto;
 import pt.estga.monument.dtos.MonumentRequestDto;
