+++ monument/src/main/java/pt/estga/monument/controllers/MonumentController.java
@@ -9,7 +9,7 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pt.estga.monument.MonumentMapper;
+import pt.estga.monument.mappers.MonumentMapper;
 import pt.estga.monument.dtos.MonumentFilter;
 import pt.estga.monument.services.MonumentService;
 import pt.estga.monument.dtos.MonumentDto;
