+++ monument/src/main/java/pt/estga/monument/services/MonumentService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pt.estga.monument.dtos.MonumentFilter;
 import pt.estga.monument.entities.Monument;
-import pt.estga.monument.MonumentRepository;
+import pt.estga.monument.repositories.MonumentRepository;
 import pt.estga.commoninfra.jpa.SpecBuilder;
 import pt.estga.commonweb.exceptions.ResourceNotFoundException;
