+++ monument/src/test/java/pt/estga/monument/services/MonumentServiceTest.java
@@ -9,7 +9,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import pt.estga.monument.entities.Monument;
-import pt.estga.monument.MonumentRepository;
+import pt.estga.monument.repositories.MonumentRepository;
 import pt.estga.commonweb.exceptions.ResourceNotFoundException;
 
 import java.util.List;
